Make use of ignoreWarnings option
This allows destructive deployment of components without worrying about
whether the component actually exists in the target org.

This is especially helpful for when destructive changes need to be
checked in to version control, as developers can safely reuse the same
json file even if they have used a previous version of it already.